+++ Cargo.toml
@@ -15,12 +15,10 @@ version     = "0.12.0"
 array-init      = "2.1.0"
 bytemuck        = "1.23.1"
 crossbeam-queue = "0.3.2"
-loom            = "0.7.2"
 thiserror       = "2.0.12"
 
 [dev-dependencies]
 bincode = { version = "2.0.1", features = ["derive"] }
-loom    = "0.7.2"
 
 [features]
 loom = []
